+++ src/interfaces/IQuorumManager.sol
@@ -6,10 +6,10 @@ interface IQuorumManager {
     /// @param srcChainId_ is the chain id from which the message (payload) is sent
     /// @param quorum_ the minimum number of message bridges required for processing
     /// NOTE: overriding child contracts should handle the sender validation & setting of message quorum
-    function setRequiredMessagingQuorum(uint16 srcChainId_, uint256 quorum_) external;
+    function setRequiredMessagingQuorum(uint64 srcChainId_, uint256 quorum_) external;
 
     /// @dev returns the required quorum for the srcChain & dstChain
     /// @param srcChainId_ is the chain id from which the message (payload) is sent
     /// @return quorum_ the minimum number of message bridges required for processing
-    function getRequiredMessagingQuorum(uint16 srcChainId_) external view returns (uint256 quorum_);
+    function getRequiredMessagingQuorum(uint64 srcChainId_) external view returns (uint256 quorum_);
 }
